#new feature for Room management

+++ API/Controllers/UserController.cs
@@ -37,7 +37,7 @@ namespace BookingApp.Controllers
 
 
         [HttpDelete]
-        public async Task<ActionResult> DeleteAsync(decimal id)
+        public async Task<ActionResult> DeleteAsync(int id)
         {
             return StatusCodeResult(await _service.DeleteAsync(id));
         }
